Hid unlicensed plugins in the restricted environment. Made it so some plugins can have their uninstall status changed. This can be used when they are all compiled together on ios.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@5461 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ AnomalousMedical/GUI/DownloadManagerGUI/DownloadManagerGUI.cs
@@ -108,7 +108,7 @@ namespace Medical.GUI
                     }
                     foreach (AtlasPlugin plugin in pluginManager.UnlicensedPlugins)
                     {
-                        if (plugin.AllowUninstall)
+                        if (plugin.AllowUninstall && PlatformConfig.UnrestrictedEnvironment)
                         {
                             plugin.loadGUIResources();
                             addInfoToButtonGrid(new UninstallInfo(plugin, ServerDownloadStatus.Unlicensed), false);
